Move ReportRow into Audits module
We may in future move this into domain alongside Report class.
For now we are making this initial change and will revisit
in future.
