+++ app/models/link.rb
@@ -7,7 +7,7 @@ class Link < ApplicationRecord
   TOPICS = "topics".freeze
   TAXONOMIES = "taxons".freeze
 
-  after_save { ReportRow.precompute(source) }
+  after_save { Audits::ReportRow.precompute(source) }
 
   belongs_to :source,
     class_name: :ContentItem,
