+++ spec/models/audits/audit_spec.rb
@@ -34,8 +34,8 @@ RSpec.describe Audits::Audit do
 
   describe "callbacks" do
     it "precomputes the content_item's report row after saving" do
-      expect { subject.save! }.to change(ReportRow, :count).by(1)
-      expect { subject.save! }.not_to change(ReportRow, :count)
+      expect { subject.save! }.to change(Audits::ReportRow, :count).by(1)
+      expect { subject.save! }.not_to change(Audits::ReportRow, :count)
     end
   end
